Add category fallback and Showroom component
